First set to the rework of the Update System
I added multiple thing to Update.py like adding functions to add like small modules. I also added a system to verify if the file is already prensent to sort of "save" badwitch ( even if the manifest is like 3 bytes ). This is the first step of a updated update system so the end user don't have to download every new version from github ! Next step, adding stuff to the manifest so he knows where to pull the files from !