Update 0.2.0 (#6)
* Update V0.0.2

* Update V0.0.3

* Update manifest.json

Test Manifest

* Update README.md

* V0.0.3

* Updated the URL with the public repo

* Added the rename thing and the remove thing to have just one manifest :)

* Previous commit, same comment

* Added a little popup ( WIP, subject to changes )

I added a little popup if the Update.py detect a newer JSON. For the next update I will need to add a system to make the popup display the Updates tab. And then a complete Update system with a bigget JSON file !

* First set to the rework of the Update System

I added multiple thing to Update.py like adding functions to add like small modules. I also added a system to verify if the file is already prensent to sort of "save" badwitch ( even if the manifest is like 3 bytes ). This is the first step of a updated update system so the end user don't have to download every new version from github ! Next step, adding stuff to the manifest so he knows where to pull the files from !

* Translation

I translated all files, this took me a day :/

* Cleaned a little bit

* Resolve merge conflict by incorporating both suggestions